requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.39.0
jinja2==3.1.4
networkx==3.3
pandas==2.2.2
chromadb==0.5.4
sentence-transformers==3.0.1
scikit-learn==1.5.1
reportlab==4.2.2
requests==2.32.3
